size_t type updates.

+++ code/SkeletonMeshBuilder.cpp
@@ -84,7 +84,7 @@ SkeletonMeshBuilder::SkeletonMeshBuilder( aiScene* pScene, aiNode* root, bool bK
 void SkeletonMeshBuilder::CreateGeometry( const aiNode* pNode)
 {
 	// add a joint entry for the node. 
-	const unsigned int vertexStartIndex = mVertices.size();
+	const size_t vertexStartIndex = mVertices.size();
 
 	// now build the geometry. 
 	if( pNode->mNumChildren > 0 && !mKnobsOnly)
@@ -169,7 +169,7 @@ void SkeletonMeshBuilder::CreateGeometry( const aiNode* pNode)
 		mFaces.push_back( Face( vertexStartIndex + 21, vertexStartIndex + 22, vertexStartIndex + 23));
 	}
 
-	unsigned int numVertices = mVertices.size() - vertexStartIndex;
+	size_t numVertices = mVertices.size() - vertexStartIndex;
 	if( numVertices > 0)
 	{
 		// create a bone affecting all the newly created vertices
@@ -185,18 +185,18 @@ void SkeletonMeshBuilder::CreateGeometry( const aiNode* pNode)
 		// add all the vertices to the bone's influences
 		bone->mNumWeights = numVertices;
 		bone->mWeights = new aiVertexWeight[numVertices];
-		for( unsigned int a = 0; a < numVertices; a++)
+		for( size_t a = 0; a < numVertices; a++)
 			bone->mWeights[a] = aiVertexWeight( vertexStartIndex + a, 1.0f);
 
 		// HACK: (thom) transform all vertices to the bone's local space. Should be done before adding
 		// them to the array, but I'm tired now and I'm annoyed.
 		aiMatrix4x4 boneToMeshTransform = aiMatrix4x4( bone->mOffsetMatrix).Inverse();
-		for( unsigned int a = vertexStartIndex; a < mVertices.size(); a++)
+		for( size_t a = vertexStartIndex; a < mVertices.size(); a++)
 			mVertices[a] = boneToMeshTransform * mVertices[a];
 	}
 
 	// and finally recurse into the children list
-	for( unsigned int a = 0; a < pNode->mNumChildren; a++)
+	for( size_t a = 0; a < pNode->mNumChildren; a++)
 		CreateGeometry( pNode->mChildren[a]);
 }
